requirements.txt
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py37_0
asn1crypto=1.3.0=py37_0
astor=0.8.0=py37_0
astroid=2.3.3=py37_0
blas=1.0=mkl
blinker=1.4=py37_0
ca-certificates=2020.1.1=0
cachetools=3.1.1=py_0
certifi=2020.4.5.1=py37_0
cffi=1.14.0=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.0=py37_0
colorama=0.4.3=py_0
cryptography=2.8=py37h7a1dbc1_0
cudatoolkit=10.1.243=h74a9793_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py37_0
freetype=2.9.1=ha9979f8_1
gast=0.2.2=py37_0
google-auth=1.11.2=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.1.8=py_0
grpcio=1.27.2=py37h351948d_0
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=py37_0
intel-openmp=2020.0=166
isort=4.3.21=py37_0
joblib=0.14.1=py_0
jpeg=9b=hb83a4c4_2
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37ha925a31_0
lazy-object-proxy=1.4.3=py37he774522_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.11.4=h7bd577a_0
libtiff=4.1.0=h56a325e_0
llvmlite=0.31.0=py37ha925a31_0
markdown=3.1.1=py37_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37h64f37c6_0
mccabe=0.6.1=py37_1
mkl=2020.0=166
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
numba=0.48.0=py37h47e9c7a_0
numpy=1.18.1=py37h93ca92e_0
numpy-base=1.18.1=py37hc3f5095_1
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
openssl=1.1.1f=he774522_0
opt_einsum=3.1.0=py_0
pillow=7.0.0=py37hcc1f983_0
pip=20.0.2=py37_1
protobuf=3.11.4=py37h33f27b4_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.19=py37_0
pyjwt=1.7.1=py37_0
pylint=2.4.4=py37_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pysocks=1.7.1=py37_0
python=3.7.6=h60c2a47_2
python-dateutil=2.8.1=py_0
qt=5.9.7=vc14h73c81de_0
requests=2.23.0=py37_0
requests-oauthlib=1.3.0=py_0
rsa=4.0=py_0
scikit-learn=0.22.1=py37h6288b17_0
scipy=1.4.1=py37h9439919_0
setuptools=46.0.0=py37_0
sip=4.19.8=py37h6538335_0
six=1.14.0=py37_0
sqlite=3.31.1=he774522_0
tbb=2020.0=h74a9793_0
tensorboard=2.1.0=py3_0
tensorflow=2.1.0=gpu_py37h7db9008_0
tensorflow-base=2.1.0=gpu_py37h55f5790_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
tensorflow-gpu=2.1.0=h0d30ee6_0
termcolor=1.1.0=py37_1
tk=8.6.8=hfa6e2cd_0
tornado=6.0.4=py37he774522_1
urllib3=1.25.8=py37_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
werkzeug=0.16.1=py_0
wheel=0.34.2=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
wrapt=1.11.2=py37he774522_0
xz=5.2.4=h2fa13f4_4
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0